+++ generator9000_app/app/components/export_modal.tsx
@@ -35,7 +35,7 @@ const ExportModalComponent: React.FC<ExportModalComponentProps> = ({ includeImag
                 </div>
                 <div className="form-control my-4">
                     <label className="label">
-                        <span className="label-text">Select Bucket:</span>
+                        <span className="label-text">Select Image Bucket (Coming soon)...</span>
                     </label>
                     <select disabled={true} value={selectedBucket} onChange={handleDropdownChange} className="select select-primary w-full max-w-xs">
                         <option>No Bucket</option>