+++ generator9000_app/app/components/prompt_menu.tsx
@@ -113,23 +113,18 @@ const PromptMenuComponent: React.FC<PromptMenuComponentProps> = ({ prompt, image
                         ))}
                     </select>
                 </div>
-
                 <p className='text-xs font-light mb-1'>Data Prompt</p>
+                <label className="label cursor-pointer">
+                    <p className="label-text text-xs opacity-50">Enable Data Generation</p>
+                    <input type="checkbox" checked={generateData} onChange={(e) => { setGenerateData(e.target.checked); setGenerateOptions(e.target.checked, generateImage) }} className="checkbox checkbox-sm" />
+                </label>
                 <textarea disabled={!generateData} className="textarea textarea-bordered w-full" placeholder="Enter your prompt here" value={prompt} onChange={(e) => setPrompt(e.target.value)}></textarea>
-                <div className="form-control">
-                    <label className="label cursor-pointer">
-                        <span className="label-text">Generate Data</span>
-                        <input type="checkbox" checked={generateData} onChange={(e) => { setGenerateData(e.target.checked); setGenerateOptions(e.target.checked, generateImage) }} className="checkbox" />
-                    </label>
-                </div>
-                <p className='text-xs font-light mb-1'>Image Prompt</p>
+                <p className='text-xs font-light mb-1 mt-4'>Image Prompt</p>
+                <label className="label cursor-pointer">
+                    <p className="label-text text-xs opacity-50">Enable Image Generation</p>
+                    <input type="checkbox" checked={generateImage} onChange={(e) => { setGenerateImage(e.target.checked); setGenerateOptions(generateData, e.target.checked) }} className="checkbox checkbox-sm" />
+                </label>
                 <textarea disabled={!generateImage} className="textarea textarea-bordered w-full" placeholder="Enter your image prompt here" value={imagePrompt} onChange={(e) => setImagePrompt(e.target.value)}></textarea>
-                <div className="form-control">
-                    <label className="label cursor-pointer">
-                        <span className="label-text">Generate Image</span>
-                        <input type="checkbox" checked={generateImage} onChange={(e) => { setGenerateImage(e.target.checked); setGenerateOptions(generateData, e.target.checked) }} className="checkbox" />
-                    </label>
-                </div>
 
             </div>
             <div className='p-2'>