+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# v0.2.0 (Open Source Release)
+
+## Added
+- Clear API Key
+
+## Fixed
+
+## Changed
+- UI Changes for improved UX
+    - Fix colors
+    - Transform settings into dropdown
+    - File Input as Button
+    - Make everything a bit smaller
+
 # v0.1.0 (Open Source Release)
 
 ## Added